feat: connect classifier service to the router
